Define HttpErrorService
